+++ Editor/AvatarBuildHook.cs
@@ -1,5 +1,4 @@
 ﻿#if UNITY_EDITOR
-#if HAS_IEDITOR_ONLY
 using System;
 using UnityEngine;
 using UnityEditor;
@@ -65,5 +64,4 @@ namespace d4rkpl4y3r.AvatarOptimizer
         }
     }
 }
-#endif
 #endif